Remove check for metainfo.its
It's not needed as detected ${METAINFO_ITS} is not used anywhere, and it breaks the build on e.g. FreeBSD where the file resides in e.g. `/usr/local/share/gettext-0.21/its` which is neither detected correctly nor can be specified manually due to unknown gettext version on the target system.